Log on materializing GlobalsArray
Summary:
"The time has come," the HAM team said,
"To kill GlobalsArray.
Its reffiness offends the eyes
Since COW is here to stay,
And it implements Vsize
Which slows down CountArray."

"But wait a bit," Hack users cried,
"Will our code still be right?
For someone may, for Good Reason
To $GLOBALS read or write -
And if the key were...'GLOBALS'
It would take down the site!"

"A bunch of clowns!" the HAM team said,
And heaved a weary sigh.
"These notices in lookupAdd
Will fire if aught's awry.
If they've the gall to read 'em all
We'll warn in getDefined!"

The warnings shipped with next release.
The error logs looked clean.
Soon enough we raised instead
And still the push was green -
For when they tried to file a SEV:
"Response code: 418"

Reviewed By: mxw

Differential Revision: D21705642

fbshipit-source-id: 1bf845d1b45292fe3ed857243ffc9f15e03d52e7