Note invariants in HHBBC type-system
Summary:
I need to modify HHBBC to handle member instructions post HAM specialization correctly. In the course of doing so, I noticed that our logic here computes "isVector" as "could-be", but it's used in some places as "could-be" and in other places as "must-be". I was confused!

It turns out that at this location, these two conditions are identical. The proof requires non-local reasoning. In union_of, we decline to represent unions like Vec|Dict, pessimizing them to InitCell. Then, when doing member instruction logic, if we don't know that the base is array-like, we fall back to generic logic. Hence, at the points we're running this logic, the base "could-be" a Vec iff it "must-be" one.

Confirm this reading with assertions and land it.

Reviewed By: dneiter

Differential Revision: D21712375

fbshipit-source-id: b2de261c05d2b3cd6a2b94706a2f381623705500