Fix PlainArray IR unsoundness
Summary:
NewMixedArray, NewStructArray, and AllocStructArray now produce PlainArray, but are still typed as returning DArrMixed. Have them return D(Arr) instead. This change is necessary for the JIT type system to be sound. It has no effect on production behavior, since I don't see the plain array bytecodes used anywhere in prod.

It's awkward to have an IR op called NewMixedArray that returns a PlainArray, so rename to NewPlainArray.

It's also awkward that we have to widen the source type of InitMixedLayoutArray here, but I am not going to introduce TPlainArr to the JIT. Instead, we'll wait until billf bans array(...) syntax and simply delete the PlainArray ops. Widening the source type only affects debug mode checks - and again, these bytecodes aren't used.

As a sanity check that we've caught everything, we note that the DArrMixed and DArrPacked dest types are now dead! Delete them.

Reviewed By: dneiter

Differential Revision: D21717100

fbshipit-source-id: 5fed3b07d7174a565155a9e5f9f52784966925ce