Drop dvarray jit::Type specialization
Summary:
Now that ArrayKind implies dvarray-ness, we can drop the "dvarray" bit in the JIT's ArraySpec. This change is a small pessimization in places where we union varrays and darrays, but it's not common enough to matter. (We still elide most CheckDVArray bits despite this pessimization, because we usually narrow to either varray or darray prior to doing that check.)

After this change, TVArr and TDArr mean "the type of (possibly bespoke) varrays and darrays". TPackedArr == TVArr & TVanillaArr, and TMixedArr == TDArr & TVanillaArr.

Reviewed By: ricklavoie

Differential Revision: D21717947

fbshipit-source-id: a71324fbbb9c189dc45521a9fc2d6fe48d2a476b