Fix bug in HHBBC array member ops
Summary: In the middle of a member op sequence, we may detect that we're definitely going to throw. We handle this case for Hack arrays by recursively returning TBottom as the value of each intermediate step; when we mark the base value as TBottom, we'll mark the rest of the block unreachable. Do so for PHP arrays as well.

Reviewed By: ricklavoie

Differential Revision: D21732347

fbshipit-source-id: b0e7f96f7a810cacba586e25b4819f1028fc5a9e