Fix bug in HHBBC ProvTag usage
Summary:
Prior to specialization, the main way we got a union of multiple empty arrays is due to them having different provenance tags. Post specialization, we also hit this case for a union of varray[] and darray[]. That reveals a bug here - we need to check if EvalArrayProvenance is on prior to looking up the tag.

We already have to do so in a few other places, so pull the logic into a helper.

Reviewed By: ricklavoie

Differential Revision: D21732988

fbshipit-source-id: 0da858efd7c2c13adc56f0b024ab1afd9fd9d235