Pre-emptively fix/delete tests implicitly casting non-arraykey keys for arrays 4/n
Summary:
D20231771 for context. Just plugging away at some more tests.

This test added in D9662556 appears to be testing how backtraces handle inlining. It does `hphp_array_idx(varray[], false)` to trigger a notice, but that's going to become an exception, so we need to replace `hphp_array_idx` with something else.

Differential Revision: D21598664

fbshipit-source-id: e787b17b2613a0aceb648b987285d5e17dd971c2