Fix HHBBC analysis of builtin properties
Summary:
HHBBC assumes that any property of a builtin class can be
mutated by the runtime in an unseen way. However, it is assumed that
the runtime will not violate any declared type-hints while doing
so. So, in this situation we widen the inferred type of a property to
at least include its type-hint.

However, we only do this widening after we've done the class
analysis. This means that by the time we've widened the types, we've
already analyzed functions with the narrower types, which aren't
correct. This can cause incorrect analysis, as a function (for
example) which just returns a property will have a narrower type than
the property itself.

Fix this by doing the widening before class analysis, not after.

Differential Revision: D21738609

fbshipit-source-id: eab62cdfe59f6f8b09a5a9349c0ce74c8ed1b534