Move check for missing type hint on class constants from decl into typing
Summary:
We require type hints on global constants, and require type hints on class constants if there is no initializer or if the initializer is not a certain  form (essentially, literal constants).

But: global constants are checked twice, during the decl phase, and during a NAST check. Class constants are checked during the decl phase. We don't want to emit errors during decl, and it produces some "bad position" errors in a test when positions are checked on a definition-level granularity, so this diff removes the checks from the decl phase, and adds the class constant check to the typing phase instead.

We also change slightly the set of expressions permitted as initializers in the absence of types, to include `null`, and to allow only `-` and `+` unary operators. This makes it easier to  specify: we can simply say "if no type is specified, then the initializer must be a literal".

Future: it's a bit odd that decl actually infers types anyway, and is not consistent with other declarations in Hack (methods, functions, properties, etc.) where explicit types are required. We might choose to require types on class constants; alternatively we might just add them to the AST during lowering.

Differential Revision: D21725335

fbshipit-source-id: eac2c3f4a764e9087f83e98f4ac2d41d79e90481