Eliminate dvarray aux bits
Summary:
Now that ArrayKind implies dvarray-ness, we can drop the aux bits that we used to track it before.

To do so, we just have to reimplement CheckVArray, CheckDArray, and CheckDVArray given the new representation. Ignoring bespoke types, CheckVArray and CheckDArray are simple tests for Packed and Mixed kind, and CheckDVArray is equivalent to kind() <= kMixedKind. I'll follow up next with fixes for bespoke types.

`hp --count 2 --dbgo && hp --count 4`
Not creating diff for commit cloud rev '.'
7a6527c512d74734efb74bd4296ae2866ee0fe12 exists in origin; not creating diff.
https://our.intern.facebook.com/intern/servicelab/41209826/
https://our.intern.facebook.com/intern/servicelab/41209831/
Not creating diff for commit cloud rev '.'
7a6527c512d74734efb74bd4296ae2866ee0fe12 exists in origin; not creating diff.
https://our.intern.facebook.com/intern/servicelab/41209862/
https://our.intern.facebook.com/intern/servicelab/41209861/
https://our.intern.facebook.com/intern/servicelab/41209863/
https://our.intern.facebook.com/intern/servicelab/41209864/

Reviewed By: ricklavoie

Differential Revision: D21751926

fbshipit-source-id: df79df8037ffb51bcab02a7eb89e351e8772d480