Assert tvIsPlausible before logging typehint violation notice
Summary:
We suspect, that some of unexplainable rare typehint violation notices are caused by heap corruption.

To help validating it, let's assert that tvIsPlausible() - it does enough sanity-checking, that it's likely to catch any sort of garbage being interpreted as hack value.

describe_actual_type() is only used for assembling various error messages, so there shouldn't be much overhead from doing a bit of extra validation there.

And if we already hit some sort of corruption, HHVM is very likely to crash anyway, so might as well do it earlier and with more useful error.

Reviewed By: ricklavoie

Differential Revision: D21752538

fbshipit-source-id: bb25f06318ee80a6fdc7fbf064173cf52434b632