Add position to Tany types for functions missing a return typehint
Summary:
When a function type is missing a return typehint, OCaml decl gives the resulting Tany type an inference Reason pointing to the name of the function. We should do the same in the direct decl parser.

This diff also does a bit of cleanup, replacing a few instances where we manually construct `Ty_::Tany` with a constant.

Reviewed By: Wilfred

Differential Revision: D21464071

fbshipit-source-id: bf50220bb8fd7fe5a060bd0ec9ae1efa3ae69f74