Unescape string literals
Summary:
The contents of string literals are reflected in decls, in class consts and __Deprecated messages. This means we need to handle escape sequences. We also need to be more careful removing the surrounding quotes--trim_start_matches and trim_end_matches would remove too many quotes in literals like `SINGLE_ENDS_IN_SINGLE` in the added test case. Reuse the logic from the lowerer for both escaping and unquoting.

Because it was convenient to do after some related refactoring, this diff now has the allocated decls borrow identifiers from the source text (when possible) instead of copying them into the arena. This means that the arena and the decls it contains must not outlive the source text, but that requirement was already enforced by the lifetimes declared by the direct decl parser anyway, so we might as well make use of it. If we'd like to have the ability to free the source text while its decls exist, we'll need to do some refactoring to declare that the source text's lifetime and the arena's lifetime are unrelated.

Reviewed By: Wilfred

Differential Revision: D21468333

fbshipit-source-id: d49284acba8a5c00c38d57ad84b4686ccd67a1bb