Handle type parameters in constraints
Summary:
The types specified in a constraint on a type parameter may reference other type parameters from the same parameter list, including type parameters which have not yet been parsed:

```
interface I<
  Ta as C<Tb>,
  Tb,
>{}
```

`C<Tb>` should be represented as `Tapply("\C", [Tgeneric("Tb")])`. However, as currently implemented, the direct decl parser produces `Tapply("\C", [Tapply("\Tb", [])])`. This is because in the smart constructor for `C<Tb>` (`make_generic_type_specifier`), there is no way for the direct decl parser to know that the type parameter `Tb` is in scope--the lexer has not advanced that far yet.

If we want to handle this correctly, but still build types eagerly (avoiding an IR or reallocation for types *outside* constraints), I think we must reallocate the type of each constraint in `make_type_parameters`, rewriting `Tapply`s referring to type parameters to `Tgeneric`, and this diff does so.

Reviewed By: Wilfred

Differential Revision: D21481119

fbshipit-source-id: 5589a7f235d334274cb0f591e7b8907e34ae7fb9