Allow for smooth "hot weight" interpolation
Summary:
As soon as we set the "UseHotWeight" vasm-layout runtime option, we move a large part (around 40%) of the code that's currently hot into frozen. This discrete change in behavior is hard to tune in prod - this parameter needs to be tuned along with other continuous parameters (like the TC size parameters and the hot and cold thresholds), and by making it a flag flip, we basically need to understand the new regime for these other parameters all at once.

Instead, we relax this parameter to a continuous one: a "HotWeightMultiplier" option, such that the base weight for a given translation is max(entry, multiplier * hotWeight)). When the multiplier is 0, we recover current behavior, and when it's 1, we recover the behavior with the flag on, but now we can also interpolate.

Reviewed By: ricklavoie

Differential Revision: D21764769

fbshipit-source-id: 120690424b879e4cc522c22774c107a4dce235ab