Add BespokeDVArray and fix typechecks
Summary:
Now that ArrayKind implies dvarrayness, we need two new kinds for bespoke varrays and darrays.

Add HeaderKind::BespokeVArray and HeaderKind::BespokeDArray. After adding these new kinds, we need to make sure that we can still implement all of the array-like type predicates efficiently. We do so by arranging kinds so that a bespoke kind is always next to its associated vanilla kind:
- To test isVanilla, we check that the low bit is unset.
- To test isPHPArrayType and isHackArrayType, we do comparisons.
- To test isDArray, we mask off the low bit and compare to kMixedKind; isVecType, isDictType, etc. are similar.
- As a special case, isVArray and isDVArray can always be done without masking.
- As a special case, CheckDArray can be done without masking if we know the input is vanilla.

In most cases, the new code is roughly equivalent to the old. A few are improved - isPHPArrayType and isHackArrayType are now just comparisons, and the other is...Type helpers are simplified. The new code for, say, isDictType is easier to read, and it handles bespoke types correctly as well.

Reviewed By: ricklavoie

Differential Revision: D21759267

fbshipit-source-id: b70914e5ec3e36859813e389ecd19c7354bd1aff