Remove ToOxidized trait
Summary:
I'd like to remove typechecker types from the oxidized crate and replace them with generated oxidized_by_ref definitions. This would allow deleting typechecker types and decls from the oxidized crate, since they are used only in this debugging-only path. It would also reduce the amount of conversion necessary for `to_oxidized` (since we need not convert the oxidized_by_ref structures the typechecker already uses, like `Ty`).

This diff removes the ToOxidized trait. Some of the types which implement it require an arena to allocate their oxidized_by_ref equivalent, and it's a bit easier to thread the arena through only the inherent methods which need it rather than adding it to the signature of the trait method `ToOxidized::to_oxidized`.

Reviewed By: Wilfred

Differential Revision: D21587515

fbshipit-source-id: c21b9436ba206cd8d9ec89b1fca680014a085a7b