Include typechecker types
Summary:
For each typing file currently included in the oxidized crate, also produce an oxidized_by_ref version.

This required implementing a few std traits on the Map and Set in arena_collections, since oxidized types derive Ord/PartialOrd/Eq/PartialEq/Hash by default (a few oxidized types also derive Default).

I replaced an `additional_derives` list in hh_oxidize with a bool parameter for `force_derive_copy` to make the logic a bit cleaner (since Copy was the only additional derive we ever included in that list). The idea behind the name is that by default, types derive copy if they appear in the `derive_copy` list, with the exception of C-like enums, for which we force-derive Copy.

Reviewed By: Wilfred

Differential Revision: D21587513

fbshipit-source-id: 7ca33190fd961e0097bc05ebff6c4f136d2f3dd1