Switch most maps/sets to arena_collections::{Map, Set}
Summary: In decls and types, I think arena_collections::{SortedAssocList, SortedSet} are appropriate, since they are compact, cache-friendly, and offer log(N) lookups. However, they don't provide functional-update methods. This is fine for decls and types, which we don't mutate, but it makes them inappropriate for the approach we are using in the typechecker (which is maintaining the functional style from the OCaml implementation), where we need an inexpensive means of updating immutable maps (e.g., substitution maps) in the typechecker env. This diff switches to using the tree-based Maps and Sets for maps and sets outside of decls and types.

Reviewed By: Wilfred

Differential Revision: D21587517

fbshipit-source-id: a9750fe95458b4fcbf618707175c0d330324eacd