Recursive type errors should not be obscured by trivial type aliases
Summary: Recursive types through trivial aliases weren't reported as errors. This fixes it, as well as simplifying the logic associated with detecting cycles. The list of expanded type names in `type_expansions` now includes a boolean that indicates whether or not to report a cycle as an error. Such cycles should be reported only if the original "root" of the type expansion was in checking a type definition or type constant.

Reviewed By: vsiles

Differential Revision: D21720539

fbshipit-source-id: b203e5cda7a6286df1aadaee0527c154c4580e0b