Fix unexpected typedef naming error
Summary:
```
type IntTypedef = int;

function test(): void {
  (int)((IntTypedef $amount) ==> 0);
}
```

was previously throwing an unexpected typedef error at the IntTypedef position in the lambda parameter.

Although the `Cast` node does not allow the hint to be a typedef, it should allow the expression that is being casted to include typedefs.

Rather than setting the environment correctly in naming, we rely on the invariant that the parser only allows for certain hints to be cast to. I believe they are like `string | int | bool`

I think that this mistake is another argument against the visitor pattern we have been employing for our checks :/

Reviewed By: arxanas

Differential Revision: D21782094

fbshipit-source-id: 1583f87f80486dffb1906b0cb08b1f524ca902ba