Create `NamingHash` module
Summary:
Trying again for {D20169351}, as it caused a performance regression last time. Differences:

- I've split up the FFI interface into two functions (`hash1_ocaml` and `hash2_ocaml`) rather than a single function that accepts a complete OCaml variant (`hash_variant`). This didn't appear to have any impact on performance, but is a bit better for type-safety.
- Changed from using `std::collections::hash_map::DefaultHasher` to `fnv::FnvHasher`.

Original summary:

This module is responsible for unifying the naming and dependency hashes. We use a smaller hash for dependency hashes, and a bigger hash for naming hashes, in such a way that we can attempt to query one from the other. See comments on the module.

Differential Revision: D21184301

fbshipit-source-id: 14791edf2964c6962a2016af091d60cb9fd5d34e