Better visibility, cleanup behavior for run.php's tempdir.
Summary:
Put the test run's temporary directory in a place that's easy to find with ls. On failure, when running interactively, put the test-failures and tests-passed files in the tempdir; this means the existing failure output shows paths in the tempdir for easy debugging.

Always remove the tempdir at the end if running from testpilot or when recording failures; never remove it if `--no-clean` was passed; otherwise remove it if all the tests passed. Correctly register for tempdir cleanup to happen even if we're running just 1 test and didn't fork. We were registering a function that takes no args as a signal handler, which caused warnings about it receiving too many args; fix this.

Reviewed By: fredemmott

Differential Revision: D21522293

fbshipit-source-id: de4f2975adaa6f48cf82998feb841f03782ce755