Move .out and .diff files out of the source tree.
Summary:
Move .out and .diff files out of the source tree and into the run's tempdir. Make and use helper functions for the many places that write diffs to or fetch diffs back from disk.

Add a command line option `--write-to-checkout` that opts out and leaves .out and .diff files where they currently are (next to the test). This does not play nice with eden but will allow people to avoid changing their test running workflows.

Reviewed By: oulgen

Differential Revision: D21570760

fbshipit-source-id: 753ea03b5634f3c2597f855e839ea1931df5772e