Support rematerialization involving physical registers in vasm-graph-color
Summary:
Up until now, vasm-graph-color only supported rematerializing
instructions which only used non-physical Vregs. The reason being,
there's no guarantee that a physical register has the same value at
the rematerialization point as it did when the reloaded Vreg was
defined. This is not a problem, of course, for non-physical Vregs
because of SSA.

I originally did not consider this as very important, but there's one
important use-case: Any Vreg which is defined as a lea off of the
frame pointer (commonly the stack pointer). It's beneficial to attempt
to rematerialize these doing the lea again rather than spilling it
to/from the stack.

Extend the rematerialization logic to support rematerializing
instructions which use physical registers. Most of the logic is
unchanged, except we will cache an instruction which uses physical
registers instead of caching a reload. When we attempt to use the
instruction, we need to check (along with all the usual checks) if the
involved physical registers have changed.

We want to know if any involved physical registers have changed
between where the reloaded Vreg was originally defined, and where it's
being reloaded right now. If they have not, it's safe to
rematerialize. The logic to determine this is straight-forward. Walk
backwards from the current point, visiting every predecessor. At every
block, examine every instruction (backwards), and look for an
instruction which writes to one of the physical registers *before* we
encounter the original instruction which defined the reloaded Vreg. If
we encounter no such condition along all such paths, the physical
register is safe to use.

We employ a number of tricks to speed this up. First, during liveness
analysis, we calculate some metadata about where physical registers
are written to, which lets us prune off search paths
entirely. Furthermore, we cache results at every block, to make
subsequent lookups faster. Since we can encounter loops during the
search, we need to maintain a visited set to avoid revisiting blocks.

Reviewed By: paulbiss

Differential Revision: D21423411

fbshipit-source-id: d8d1ae3c3ae46aad1ba26c5d7cc9f98e959db28c