Restore perturbation logic for color selection in vasm-graph-color
Summary:
D21146715 added logic to prefer non-REX prefix registers during color
selection, but dropped the perturbation logic to avoid ties.

This logic is useful: In situations where many registers have the same
weight (which can happen with clusters of Vregs which don't have any
preferences), the selection logic will just starting preferring the
smallest register (usually %rax) for every Vreg, which puts more
pressure on the shuffling logic (to undo bad previous choices). The
perturbation logic serves to helps this by avoiding "hot-spots" on any
one particular physical register.

Add an explicit check for a REX register or not. A register which
doesn't need a REX prefix will always be preferred to one which
does. However, if their REX-ness is the same, fall into the
perturbation case, as previously.

Reviewed By: ottoni, paulbiss

Differential Revision: D21415579

fbshipit-source-id: e7c82d8a4417ab95117a6561be1a8c4bc178b2d4