Support rematerialization with "recoverable" physical registers in vasm-graph-color
Summary:
The previous diff added the ability in vasm-graph-color to
rematerialize instructions involving physical registers. However it
could only do so if the involved physical registers were unchanged
completely along all paths from the reload point back to where the
reloaded Vreg was originally defined.

This is fine for the most part, but it prevents us from
rematerializing leas from the frame pointer in all situations. This is
because sometimes we'll move the frame pointer by some static amount,
then move it back (usually because of inlining). Even though the frame
pointer is being changed in predictable ways, we consider that as
"changed" and cannot rematerialize.

Add support for this type of thing. Instead of simply saying if a
physical register has changed or not, we instead track whether its
changed in a "recoverable" way. That is, even if the physical register
has changed, we know enough (statically) to recompute its original
value. Modify the physical register change detection logic to only
consider it truely changed if it changes in some unsupported
way. Right now we only support leas (simple ones, without any index or
scale), and copyish instructions. When attempting to rematerialize the
instruction, if the register has changed, but in a recoverable way, we
simply adjust the rematerialization instruction appropriately. Again,
this is only support for simple leas right now. We just modify the
displacement as necessary.

This opens up one further optimization. If the instruction which uses
the reloaded Vreg is using it as a Vptr base operand, we can "fold"
the rematerialization entirely into the Vptr (if the rematerialization
is a lea). So, instead of emitting a lea, then immediately using the
lea dst as the base of a Vptr, we just fold the lea directly into the
Vptr, and emit no new instruction at all. Folding a lea into a Vptr is
just a simple matter of changing the base and adding the
displacements. So, for the case where we're rematerializing a lea off
the frame pointer, then doing loads and stores (very common), we don't
generate any new instructions. In some cases this actually generates
better code than before spilling!

Make the spilling heuristic be aware of this, and prefer reloads we
can fold away entirely as described. Such reloads have zero weight,
since they generate nothing at all (normal rematerializations have a
weight of at least 1 because even though they don't reload, they still
generate code).

Differential Revision: D21424478

fbshipit-source-id: dd452c3c540b62a3e7d9599102a40eb99d827679