Work around merge conflict beween D21538253 and D21623084
Summary: I accidentally broke 2 tests by landing the stack containing D21538253, since it took the opposite approach to the already landed D21623084 on whether HPHP_TEST_TMPDIR should end in a / or not. Temporarily disable the invariant until I can fix the affected tests using HPHP_TEST_TMPDIR directly and re-enable it. We don't need to touch the HN copy because the affected tests have not made it on to the HN whitelist yet.

Reviewed By: fredemmott

Differential Revision: D21798065

fbshipit-source-id: 453cabbedd2627a0698c7bf559d088b4c443c0ec