Handle bespokes in const member ops
Summary:
The member ops Elem helper is used for const lookups, both for intermediate ops and for Isset and QueryM final ops. It's cold, though - the JIT usually calls ElemArr, ElemVec, ...

Update Elem and Isset to handle "bespoke" array-likes. ElemArr already handles bespokes, so we just need to add a helper for bespoke Hack arrays. That's useful, because the key coercion logic in ElemArr is quite complicated, and I didn't want to have to deal with it.

Reviewed By: ricklavoie

Differential Revision: D21812266

fbshipit-source-id: 933b1859bd7d7a0edba7dc669db6d6595e57d5d4