Remove redundant check for constraint satisfaction on extends and implements
Summary:
As it says in the title. Localization of `extends` and `implements` types already checks that they are correct wrt constraints on type parmaeters e.g. if we have `C<T as I> extends D<t>` and `D` has a constraint on its type parameter, then `t` will be checked against `I` under the assumption that `T <: I`. This is done in `NastCheck.class_`, and does not need to be repeated in `Typing_toplevel.class_def`.

This diff therefore removes a bunch of redundant errors.

Reviewed By: CatherineGasnier

Differential Revision: D21762661

fbshipit-source-id: 3f4c2ca0dd04bc552083dc6399589cca53db92f5