Free up m_pos for bespokes
Summary:
ArrayData uses 4 bytes for its m_pos field. It would be nice to be able to use it for other purposes for bespoke arrays.

m_pos is an array's "internal cursor", used for PHP functions like next(inout $array), etc. These functions are rarely used; most iteration is done by foreach loops that sensibly keep their state outside the array. If we enforce the invariant that all bespoke arrays have a pos fixed at iter_begin(), we can repurpose this field for bespoke arrays. Note that fixing to iter_begin() rather than to 0 is important, in case any bespoke layouts need support for tombstones or other non-contiguous iterator indices.

To do so, we just have to replace reads of m_pos with calls to getPosition() (which handles both the bespoke and vanilla cases now), and we have to ensure that we never write the field in setPosition(). We do the latter by escalating as needed in the implementation of these extension functions.

Reviewed By: mofarrell

Differential Revision: D21820301

fbshipit-source-id: 74182008358ae457724fa3e33806ae21e5ac5026