Enable logging arrays by flag
Summary: This diff adds a flag that causes us to actually generate LoggingArrays, as a way to test whether the runtime correctly handles bespoke array-likes. The runtime does NOT handle these arrays correctly right now: 3% of quick tests fail with the "emit" flag on. Of course, with the flags off, and with the "allow" flag on, tests will all pass.

Reviewed By: jjgriego

Differential Revision: D21807039

fbshipit-source-id: 94e5132ba11c1d2cb5535a9795bcbb6528598ef4