Handle bespokes in APCArray
Summary:
I still haven't decided how to handle persistent bespoke arrays. There are a few options, but the main question is that of "attribution" - we may reference the same static array from multiple units, but we need separate logging for those arrays to produce good bespoke layouts. (The static empty varray, darray, vec, and dict are extreme examples of this problem.)

In the meantime, handle bespoke arrays in APC by converting them to vanilla prior to storing them.

Reviewed By: ricklavoie, jjgriego

Differential Revision: D21825090

fbshipit-source-id: a3f1780de62c420b7af92ad24f44f48ff6c5ea8f