Track a reason for bespoke escalation
Summary: For now, we provide this reason as a static char array. We could use an enum, but it's a bit annoying - touching bespoke-array.h (which is the only interface file for bespokes) forces the recompile of a lot of code. If we need an enum, I'll change it later.

Reviewed By: ricklavoie

Differential Revision: D21825508

fbshipit-source-id: b2c5b8c1897192a32ec6d30ee5722917f58139ab