Handle bespokes in AwaitAllWaitHandle
Summary: This case is easy: we can implement a generic helper AwaitAllWaitHandle::fromArrLike. We still need to keep the array case around, because it (correctly) throws an error on GlobalsArray, and we might as well keep the current code for vec and dict around since iteration over the known layout is fast.

Reviewed By: ricklavoie

Differential Revision: D21833867

fbshipit-source-id: 335ee10aa50304cd4388420bc437c2a3cbd07ed6