Handle bespokes in conversions
Summary:
The generic conversion helpers (tvCastToFooInPlace) call out to layout-specific functions for Hack arrays. Thankfully, this isn't a case where we have to escalate; we can simply merge all of the array-like cases and rely on ArrayData's vtable.

These helpers are only used from the interpreter and from extensions / runtime locations; the JIT burns in specific type-to-type conversion calls and guards to vanilla before doing so, so we don't have to slow it down.

A quick note about assertion coverage: given that we now handle all array-likes in one case, we can't make any assertions about the type of the input any more. (Most ArrayKind implementations do assertions of their own on the input.) It looks like we're also losing coverage on the result (say, in the darray and varray cases), but that's not the case: we have another copy of these assertions down below the switch statement, so the deleted assertions about the result were redundant.

Reviewed By: ricklavoie

Differential Revision: D21830004

fbshipit-source-id: e760395aae28977ff9c3c09fb7240c49b80e66c8