Port `hh_fanout` to look up dependency hashes using naming-table saved-state
Summary: This means we use the `Naming_table.get_dep_set_files` API instead of `Typing_deps.get_files`. This commit also makes it so that we construct and pass in a naming-table saved-state for tests, to ensure that we're testing the right thing, which simplifies the saved-state loading code.

Reviewed By: ljw1004

Differential Revision: D20708531

fbshipit-source-id: fb46eb867676075f0870004e12d997c3def75cff