Handle bespokes in comparisons
Summary:
The generic ArrayData logic for these operations is correct for Hack arrays, too, so we just need to call it if either of the input arrays is bespoke. We add a new ArrayData::bothVanilla method that tests for this case with a single branch. The new helper is necessary to generate good code - see https://godbolt.org/z/R7MeoC.

As always, the JIT is guarding to vanilla and skipping the virtual dispatch here.

We are getting close to implementing full functionality for bespokes. The only remaining piece is mutable member ops, for which we will introduce new interface methods, for reasons that will become clear in the next diff. After that, there's a long tail of bugs to squash...

Reviewed By: jjgriego

Differential Revision: D21834519

fbshipit-source-id: d93c2e46567c2eaa5f75688113771db0a5b52a89