Do not trigger errors and then silence them when checking of parameter types are memoizable
Summary:
To check whether a parameter type is memoizable, we attempt a subtype test against `Container<#v>` where `#v` is a fresh type variable. We "catch" any errors produced. But unfortunately that's too late to prevent "bad primary position" notices getting logged in Scuba. (This is the main source of such logs in Scuba.) Better is to expose the subtype simplification logic so that instead of producing errors we determine whether the simplified proposition determined from the subtype assertion is trivially valid or not.

TODO: wrap up the whole "solve wrt fresh type variables" logic in a better way, so that clients don't have to manage the open tyvars, solving, etc.

Reviewed By: manzyuk

Differential Revision: D21818588

fbshipit-source-id: 7ff2f0550ac9467b512a1ebfb4d468f040861328