Fix autoimport behavior in namespaces
Summary:
If a name has a leading backslash, we should never consider it to be referring to an imported type (whether that type is imported via the autoimport mechanism or not).

If we are in a namespace, the autoimport behavior should still trigger for unqualified names, but it currently doesn't--we prefix a name with the current namespace before we resolve imports. This isn't right--we should resolve imports before prefixing the type name with the current namespace.

Reviewed By: Wilfred

Differential Revision: D21564155

fbshipit-source-id: 4e67a9339f1ec85e2d69af451f93ae97bacf2feb