Use DSL for test cases
Summary: I want to add the ability to calculate either fanout or typechecking errors in a test case. It's getting unwieldy to use a test repository without a description of what test we should be running. This commit introduces a simple DSL to write tests. (We'll be adding a `calculate-errors` command in the next commit.)

Reviewed By: 2BitSalute

Differential Revision: D21754107

fbshipit-source-id: deff63756d4d577a5b6d48d38ff528735203912e