Create static LoggingArrays
Summary:
We need to support bespoke counterparts of static arrays if we're going to have any reasonable coverage of the codebase. These bespokes need to be static, because HHBBC tracks and makes optimizations based on which values are static.

To do so, we add a path to create static LoggingArrays. We'd like to have a different static array for each array creation site, not just for each static array; the same static empty arrays appear in many different units, but they behave quite differently at different source locations. We add a SrcKey argument to the LoggingArray constructor and use it as appropriate. The SrcKey is all we need to key the map; we assert that the resulting LoggingArray wraps the expected array. We needed the SrcKey anyway, since we'll key our profiles on it.

This diff revealed a handful of one-liner bugs within bespoke arrays which I've also fixed:
- BespokeArray::EscalateForSort...didn't actually escalate for sort.
- LoggingArray::dequeue called pop instead of dequeue
- LoggingArray::copy recursively called itself rather than the wrapper's copy
- The correct "m_pos is trivial" constraint is that it equals iter_begin(), not 0.

There are...a lot of optimizations we could make here. Two that I think will end up being important:
1. We can shard the map to reduce contention on the write lock, and/or switch to concurrent_hash_map.
2. We can create a new LowPtrLoggingArray bespoke layout that only costs 16 bytes (using the newly-freed m_pos field).

I thought about doing these items now for fun, but it's way too early to be making those optimizations.

Reviewed By: jjgriego

Differential Revision: D21854996

fbshipit-source-id: cfe4d3f9215ae39c3a1e64d873bf036670a55dfd